requirements.txt
aiopg
pyyaml
Pillow
